SUPPORT VECTOR MACHINES (SVM)
APPLICATION OF SUPPORT VECTOR MACHINES IN IDENTIFYING FACTORS RELATED TO CD4 CELL COUNT LEVELS AMONG HIV PATIENTS:

The purpose of this extensive and highly detailed project was to use the machine learning tool SVM to assist in finding out the factors related to CD4 cell count levels among HIV patients in Kenya. The details pertaining to the analysis are captured within the main body of the document, with the R programming codes placed within the appendix section.